+++ lib/src/vector/float64x2_vector.dart
@@ -247,7 +247,11 @@ class Float64x2Vector with IterableMixin<double> implements Vector {
   Vector scalarDiv(num scalar) => this / scalar;
 
   @override
-  Vector toIntegerPower(int power) => _elementWisePow(power);
+  Vector pow(num exponent) => _elementWisePow(exponent);
+
+  @override
+  @deprecated
+  Vector toIntegerPower(int power) => pow(power);
 
   @override
   Vector abs({bool skipCaching = false}) =>
@@ -449,7 +453,7 @@ class Float64x2Vector with IterableMixin<double> implements Vector {
 
   /// Returns a vector as a result of applying to [this] element-wise raising
   /// to the integer power
-  Vector _elementWisePow(int exp) {
+  Vector _elementWisePow(num exp) {
     final source = Float64x2List(_numOfBuckets);
     for (int i = 0; i < _numOfBuckets; i++) {
       source[i] = _simdToIntPow(_innerSimdList[i], exp);